Fix Continuous Integration (#23)
* Delete npm's package-lock.json

This was breaking `yarn install --frozen-lockfile` as used in the CI
script

* Delete backups from merging that were checked in

* Update yarn.lock

* Add back missing string for 403

* Revert package.json and yarn.lock to upstream

package.json had three extra packages, removing these and reverting back
to the old lockfile fixed an issue with the Jest tests. It appears,
based on the error messages, that the issue with the tests was related
to Babel, ES modules, module loading, or something similar.

* Remove bad line in reblog_service

The changes from the local only patch masked that this call to a
Pubsubhubbub method should have been deleted, as in upstream. The reblog
object here no longer has a stream_entry method, so it was throwing
errors.